stored: register backends in plugin-registry
This patch introduces a plugin regsitry that contains all loaded
backends. Storage backends will self-register their factory either at
startup (in case the backend is statically linked into the SD) or when
the backend is loaded.

From the backend's perspective, there's no difference if it is
statically or dynamically loaded anymore.

Finally, the loading of the backends is done right after loading the
configuration, so for every device we check the presence of its backend
and fail if it is missing or cannot be loaded.
