+++ core/src/stored/backends/CMakeLists.txt
@@ -26,7 +26,7 @@ macro(add_sd_backend backend_name)
     add_library(${backend_name} MODULE)
     install(TARGETS ${backend_name} DESTINATION ${backenddir})
   else()
-    add_library(${backend_name} STATIC)
+    add_library(${backend_name} OBJECT)
     target_link_libraries(bareossd PRIVATE ${backend_name})
   endif()
   set_property(TARGET ${backend_name} PROPERTY POSITION_INDEPENDENT_CODE ON)
@@ -56,14 +56,20 @@ if(TARGET droplet)
 
 endif()
 
+add_sd_backend(bareossd-file)
 add_sd_backend(bareossd-fifo)
-target_sources(bareossd-fifo PRIVATE unix_fifo_device.cc)
-if(HAVE_DARWIN_OS)
-  target_link_libraries(bareossd-fifo bareos bareossd)
-endif()
-
 add_sd_backend(bareossd-tape)
-target_sources(bareossd-tape PRIVATE unix_tape_device.cc)
+if(HAVE_WIN32)
+  target_sources(bareossd-file PRIVATE win32_file_device.cc)
+  target_sources(bareossd-fifo PRIVATE win32_fifo_device.cc)
+  target_sources(bareossd-tape PRIVATE win32_tape_device.cc)
+
+  target_link_libraries(bareossd-file PRIVATE bareos)
+else()
+  target_sources(bareossd-file PRIVATE unix_file_device.cc)
+  target_sources(bareossd-fifo PRIVATE unix_fifo_device.cc)
+  target_sources(bareossd-tape PRIVATE unix_tape_device.cc)
+endif()
 
 if(HAVE_DYNAMIC_SD_BACKENDS)
   add_library(bareossd-gentape SHARED generic_tape_device.cc)
@@ -76,7 +82,6 @@ else()
   target_sources(bareossd-tape PRIVATE generic_tape_device.cc)
 endif()
 
-if(NOT HAVE_WIN32)
-  add_sd_backend(bareossd-file)
-  target_sources(bareossd-file PRIVATE unix_file_device.cc)
+if(HAVE_DARWIN_OS)
+  target_link_libraries(bareossd-fifo bareos bareossd)
 endif()
