+++ core/src/stored/stored_conf.cc
@@ -43,6 +43,7 @@
 #define NEED_JANSSON_NAMESPACE 1
 #include "lib/output_formatter.h"
 #include "lib/output_formatter_resource.h"
+#include "lib/plugin_registry.h"
 #include "include/auth_types.h"
 #include "include/jcr.h"
 
@@ -544,9 +545,14 @@ static void GuessMissingDeviceTypes(ConfigurationParser& my_config)
   }
 }
 
-static void CheckDeviceBackends(ConfigurationParser& my_config)
+static void CheckAndLoadDeviceBackends(ConfigurationParser& my_config)
 {
   PluginRegistry<BackendInterface>::DumpDbg();
+
+#if defined(HAVE_DYNAMIC_SD_BACKENDS)
+  auto storage_res
+      = dynamic_cast<StorageResource*>(my_config.GetNextRes(R_STORAGE, NULL));
+#endif
 
   BareosResource* p = nullptr;
 
@@ -554,8 +560,25 @@ static void CheckDeviceBackends(ConfigurationParser& my_config)
     DeviceResource* d = dynamic_cast<DeviceResource*>(p);
     if (d) {
       to_lower(d->dev_type);
-      Dmsg0(50, "device %s has dev_type %s\n", d->resource_name_,
-            d->dev_type.c_str());
+      if (!PluginRegistry<BackendInterface>::IsRegistered(d->dev_type)) {
+#if defined(HAVE_DYNAMIC_SD_BACKENDS)
+        if (!storage_res || storage_res->backend_directories.empty()) {
+          Jmsg2(nullptr, M_ERROR_TERM, 0,
+                "Backend Directory not set. Cannot load dynamic backend %s\n",
+                d->dev_type.c_str());
+        }
+        if (!LoadStorageBackend(d->dev_type,
+                                storage_res->backend_directories)) {
+          Jmsg2(nullptr, M_ERROR_TERM, 0,
+                "Could not load storage backend %s for device %s.\n",
+                d->dev_type.c_str(), d->resource_name_);
+        }
+#else
+        Jmsg2(nullptr, M_ERROR_TERM, 0,
+              "Backend %s for device %s not available.\n", d->dev_type.c_str(),
+              d->resource_name_);
+#endif
+      }
     }
   }
 }
@@ -564,6 +587,7 @@ static void ConfigReadyCallback(ConfigurationParser& my_config)
 {
   MultiplyConfiguredDevices(my_config);
   GuessMissingDeviceTypes(my_config);
+  CheckAndLoadDeviceBackends(my_config);
   CheckDropletDevices(my_config);
 }
